examples/doc-scan/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asn1==2.2.0               # via yoti
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via flask
cryptography==2.9.2       # via pyopenssl, yoti
deprecated==1.2.6         # via yoti
filetype==1.0.7           # via -r requirements.in
flask==1.1.2              # via -r requirements.in
future==0.18.2            # via yoti
idna==2.9                 # via requests
iso8601==0.1.12           # via yoti
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
protobuf==3.11.3          # via yoti
pycparser==2.20           # via cffi
pyopenssl==19.1.0         # via yoti
python-dotenv==0.13.0     # via -r requirements.in
requests==2.23.0          # via yoti
six==1.14.0               # via cryptography, protobuf, pyopenssl
urllib3==1.25.9           # via requests
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via deprecated
yoti==2.11.2              # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
